[3587] Add an extension point to add new project cards
Bug: https://github.com/eclipse-sirius/sirius-web/issues/3587
