+++ packages/sirius-web/frontend/sirius-web-application/src/views/project-browser/create-projects-area/CreateProjectArea.tsx
@@ -11,13 +11,14 @@
  *     Obeo - initial API and implementation
  *******************************************************************************/
 
+import { useComponents } from '@eclipse-sirius/sirius-components-core';
 import Typography from '@material-ui/core/Typography';
 import { makeStyles } from '@material-ui/core/styles';
 import { useState } from 'react';
 import { Redirect } from 'react-router-dom';
 import { CreateProjectAreaProps, CreateProjectAreaState } from './CreateProjectArea.types';
-import { NewProjectCard, ProjectTemplateCard, ShowAllTemplatesCard, UploadProjectCard } from './ProjectTemplateCard';
-import { ProjectTemplatesModal } from './ProjectTemplatesModal';
+import { createProjectAreaCardExtensionPoint } from './CreateProjectAreaExtensionPoints';
+import { ProjectTemplateCard } from './ProjectTemplateCard';
 import { redirectUrlFromTemplate } from './redirectUrlFromTemplate';
 import { useCreateProjectFromTemplate } from './useCreateProjectFromTemplate';
 import { useProjectTemplates } from './useProjectTemplates';
@@ -44,11 +45,14 @@ const useCreateProjectAreaStyles = makeStyles((theme) => ({
 
 export const CreateProjectArea = ({}: CreateProjectAreaProps) => {
   const classes = useCreateProjectAreaStyles();
+
+  const createProjectAreaCards = useComponents(createProjectAreaCardExtensionPoint);
+
+  const limit = Math.max(0, 6 - createProjectAreaCards.length);
   const [state, setState] = useState<CreateProjectAreaState>({
     page: 0,
-    limit: 3,
+    limit,
     runningTemplate: null,
-    modalDisplayed: null,
   });
 
   const { data } = useProjectTemplates(state.page, state.limit);
@@ -62,9 +66,6 @@ export const CreateProjectArea = ({}: CreateProjectAreaProps) => {
       setState((prevState) => ({ ...prevState, runningTemplate: template }));
     }
   };
-
-  const showAllTemplatesModal = () => setState((prevState) => ({ ...prevState, modalDisplayed: 'SHOW_ALL_TEMPLATES' }));
-  const closeModal = () => setState((prevState) => ({ ...prevState, modalDisplayed: null }));
 
   const redirectUrl: string | null = projectCreatedFromTemplate
     ? redirectUrlFromTemplate(projectCreatedFromTemplate)
@@ -89,12 +90,11 @@ export const CreateProjectArea = ({}: CreateProjectAreaProps) => {
               key={template.id}
             />
           ))}
-          <NewProjectCard />
-          <UploadProjectCard />
-          <ShowAllTemplatesCard onClick={showAllTemplatesModal} />
+          {createProjectAreaCards.map(({ Component: Card }, index) => (
+            <Card key={index} />
+          ))}
         </div>
       </div>
-      {state.modalDisplayed === 'SHOW_ALL_TEMPLATES' ? <ProjectTemplatesModal onClose={closeModal} /> : null}
     </>
   );
 };
