+++ packages/sirius-web/frontend/sirius-web-application/src/views/project-browser/create-projects-area/ProjectTemplateCard.types.ts
@@ -19,7 +19,3 @@ export interface ProjectTemplateCardProps {
   disabled: boolean;
   onCreateProject: () => void;
 }
-
-export interface ShowAllTemplatesCardProps {
-  onClick: () => void;
-}
