+++ packages/sirius-web/frontend/sirius-web-application/src/views/project-browser/create-projects-area/ProjectTemplatesModal.tsx
@@ -20,8 +20,10 @@ import Pagination from '@material-ui/lab/Pagination';
 import gql from 'graphql-tag';
 import { useEffect, useState } from 'react';
 import { Redirect } from 'react-router-dom';
-import { NewProjectCard, ProjectTemplateCard, UploadProjectCard } from './ProjectTemplateCard';
+import { NewProjectCard } from './NewProjectCard';
+import { ProjectTemplateCard } from './ProjectTemplateCard';
 import { ProjectTemplatesModalProps, ProjectTemplatesModalState } from './ProjectTemplatesModal.types';
+import { UploadProjectCard } from './UploadProjectCard';
 import { redirectUrlFromTemplate } from './redirectUrlFromTemplate';
 import { useCreateProjectFromTemplate } from './useCreateProjectFromTemplate';
 import { useProjectTemplates } from './useProjectTemplates';
